refactor(shared): isolate qwen provider metadata
