+++ src/shared/api/index.ts
@@ -26,6 +26,7 @@ export * from "./providers/bedrock"
 export * from "./providers/vertex"
 export * from "./providers/openRouter"
 export * from "./providers/gemini"
+export * from "./providers/qwen"
 
 // OpenRouter
 // https://openrouter.ai/models?order=newest&supported_parameters=tools
@@ -294,702 +295,6 @@ export const huggingFaceModels = {
 		description: "Efficient 8B parameter Llama model for general-purpose tasks.",
 	},
 } as const satisfies Record<string, ModelInfo>
-
-// Qwen
-// https://bailian.console.aliyun.com/
-// The first model in the list is used as the default model for each region
-export const internationalQwenModels = {
-	"qwen3-coder-plus": {
-		maxTokens: 65_536,
-		contextWindow: 1_000_000,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 1,
-		outputPrice: 5,
-	},
-	"qwen3-coder-480b-a35b-instruct": {
-		maxTokens: 65_536,
-		contextWindow: 204_800,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 1.5,
-		outputPrice: 7.5,
-	},
-	"qwen3-235b-a22b": {
-		maxTokens: 16_384,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 2,
-		outputPrice: 8,
-		cacheWritesPrice: 2,
-		cacheReadsPrice: 8,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 20,
-		},
-	},
-	"qwen3-32b": {
-		maxTokens: 16_384,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 2,
-		outputPrice: 8,
-		cacheWritesPrice: 2,
-		cacheReadsPrice: 8,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 20,
-		},
-	},
-	"qwen3-30b-a3b": {
-		maxTokens: 16_384,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.75,
-		outputPrice: 3,
-		cacheWritesPrice: 0.75,
-		cacheReadsPrice: 3,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 7.5,
-		},
-	},
-	"qwen3-14b": {
-		maxTokens: 8_192,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 1,
-		outputPrice: 4,
-		cacheWritesPrice: 1,
-		cacheReadsPrice: 4,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 10,
-		},
-	},
-	"qwen3-8b": {
-		maxTokens: 8_192,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.5,
-		outputPrice: 2,
-		cacheWritesPrice: 0.5,
-		cacheReadsPrice: 2,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 5,
-		},
-	},
-	"qwen3-4b": {
-		maxTokens: 8_192,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.3,
-		outputPrice: 1.2,
-		cacheWritesPrice: 0.3,
-		cacheReadsPrice: 1.2,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 3,
-		},
-	},
-	"qwen3-1.7b": {
-		maxTokens: 8_192,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.3,
-		outputPrice: 1.2,
-		cacheWritesPrice: 0.3,
-		cacheReadsPrice: 1.2,
-		thinkingConfig: {
-			maxBudget: 30_720,
-			outputPrice: 3,
-		},
-	},
-	"qwen3-0.6b": {
-		maxTokens: 8_192,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.3,
-		outputPrice: 1.2,
-		cacheWritesPrice: 0.3,
-		cacheReadsPrice: 1.2,
-		thinkingConfig: {
-			maxBudget: 30_720,
-			outputPrice: 3,
-		},
-	},
-	"qwen2.5-coder-32b-instruct": {
-		maxTokens: 8_192,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.002,
-		outputPrice: 0.006,
-		cacheWritesPrice: 0.002,
-		cacheReadsPrice: 0.006,
-	},
-	"qwen2.5-coder-14b-instruct": {
-		maxTokens: 8_192,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.002,
-		outputPrice: 0.006,
-		cacheWritesPrice: 0.002,
-		cacheReadsPrice: 0.006,
-	},
-	"qwen2.5-coder-7b-instruct": {
-		maxTokens: 8_192,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.001,
-		outputPrice: 0.002,
-		cacheWritesPrice: 0.001,
-		cacheReadsPrice: 0.002,
-	},
-	"qwen2.5-coder-3b-instruct": {
-		maxTokens: 8_192,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.0,
-		outputPrice: 0.0,
-		cacheWritesPrice: 0.0,
-		cacheReadsPrice: 0.0,
-	},
-	"qwen2.5-coder-1.5b-instruct": {
-		maxTokens: 8_192,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.0,
-		outputPrice: 0.0,
-		cacheWritesPrice: 0.0,
-		cacheReadsPrice: 0.0,
-	},
-	"qwen2.5-coder-0.5b-instruct": {
-		maxTokens: 8_192,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.0,
-		outputPrice: 0.0,
-		cacheWritesPrice: 0.0,
-		cacheReadsPrice: 0.0,
-	},
-	"qwen-coder-plus-latest": {
-		maxTokens: 129_024,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 3.5,
-		outputPrice: 7,
-		cacheWritesPrice: 3.5,
-		cacheReadsPrice: 7,
-	},
-	"qwen-plus-latest": {
-		maxTokens: 16_384,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.8,
-		outputPrice: 2,
-		cacheWritesPrice: 0.8,
-		cacheReadsPrice: 2,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 16,
-		},
-	},
-	"qwen-turbo-latest": {
-		maxTokens: 16_384,
-		contextWindow: 1_000_000,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.3,
-		outputPrice: 0.6,
-		cacheWritesPrice: 0.3,
-		cacheReadsPrice: 0.6,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 6,
-		},
-	},
-	"qwen-max-latest": {
-		maxTokens: 30_720,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 2.4,
-		outputPrice: 9.6,
-		cacheWritesPrice: 2.4,
-		cacheReadsPrice: 9.6,
-	},
-	"qwen-coder-plus": {
-		maxTokens: 129_024,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 3.5,
-		outputPrice: 7,
-		cacheWritesPrice: 3.5,
-		cacheReadsPrice: 7,
-	},
-	"qwen-plus": {
-		maxTokens: 129_024,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.8,
-		outputPrice: 2,
-		cacheWritesPrice: 0.8,
-		cacheReadsPrice: 0.2,
-	},
-	"qwen-turbo": {
-		maxTokens: 1_000_000,
-		contextWindow: 1_000_000,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.3,
-		outputPrice: 0.6,
-		cacheWritesPrice: 0.3,
-		cacheReadsPrice: 0.6,
-	},
-	"qwen-max": {
-		maxTokens: 30_720,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 2.4,
-		outputPrice: 9.6,
-		cacheWritesPrice: 2.4,
-		cacheReadsPrice: 9.6,
-	},
-	"deepseek-v3": {
-		maxTokens: 8_000,
-		contextWindow: 64_000,
-		supportsImages: false,
-		supportsPromptCache: true,
-		inputPrice: 0,
-		outputPrice: 0.28,
-		cacheWritesPrice: 0.14,
-		cacheReadsPrice: 0.014,
-	},
-	"deepseek-r1": {
-		maxTokens: 8_000,
-		contextWindow: 64_000,
-		supportsImages: false,
-		supportsPromptCache: true,
-		inputPrice: 0,
-		outputPrice: 2.19,
-		cacheWritesPrice: 0.55,
-		cacheReadsPrice: 0.14,
-	},
-	"qwen-vl-max": {
-		maxTokens: 30_720,
-		contextWindow: 32_768,
-		supportsImages: true,
-		supportsPromptCache: false,
-		inputPrice: 3,
-		outputPrice: 9,
-		cacheWritesPrice: 3,
-		cacheReadsPrice: 9,
-	},
-	"qwen-vl-max-latest": {
-		maxTokens: 129_024,
-		contextWindow: 131_072,
-		supportsImages: true,
-		supportsPromptCache: false,
-		inputPrice: 3,
-		outputPrice: 9,
-		cacheWritesPrice: 3,
-		cacheReadsPrice: 9,
-	},
-	"qwen-vl-plus": {
-		maxTokens: 6_000,
-		contextWindow: 8_000,
-		supportsImages: true,
-		supportsPromptCache: false,
-		inputPrice: 1.5,
-		outputPrice: 4.5,
-		cacheWritesPrice: 1.5,
-		cacheReadsPrice: 4.5,
-	},
-	"qwen-vl-plus-latest": {
-		maxTokens: 129_024,
-		contextWindow: 131_072,
-		supportsImages: true,
-		supportsPromptCache: false,
-		inputPrice: 1.5,
-		outputPrice: 4.5,
-		cacheWritesPrice: 1.5,
-		cacheReadsPrice: 4.5,
-	},
-} as const satisfies Record<string, ModelInfo>
-
-export const mainlandQwenModels = {
-	"qwen3-235b-a22b": {
-		maxTokens: 16_384,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 2,
-		outputPrice: 8,
-		cacheWritesPrice: 2,
-		cacheReadsPrice: 8,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 20,
-		},
-	},
-	"qwen3-32b": {
-		maxTokens: 16_384,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 2,
-		outputPrice: 8,
-		cacheWritesPrice: 2,
-		cacheReadsPrice: 8,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 20,
-		},
-	},
-	"qwen3-30b-a3b": {
-		maxTokens: 16_384,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.75,
-		outputPrice: 3,
-		cacheWritesPrice: 0.75,
-		cacheReadsPrice: 3,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 7.5,
-		},
-	},
-	"qwen3-14b": {
-		maxTokens: 8_192,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 1,
-		outputPrice: 4,
-		cacheWritesPrice: 1,
-		cacheReadsPrice: 4,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 10,
-		},
-	},
-	"qwen3-8b": {
-		maxTokens: 8_192,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.5,
-		outputPrice: 2,
-		cacheWritesPrice: 0.5,
-		cacheReadsPrice: 2,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 5,
-		},
-	},
-	"qwen3-4b": {
-		maxTokens: 8_192,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.3,
-		outputPrice: 1.2,
-		cacheWritesPrice: 0.3,
-		cacheReadsPrice: 1.2,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 3,
-		},
-	},
-	"qwen3-1.7b": {
-		maxTokens: 8_192,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.3,
-		outputPrice: 1.2,
-		cacheWritesPrice: 0.3,
-		cacheReadsPrice: 1.2,
-		thinkingConfig: {
-			maxBudget: 30_720,
-			outputPrice: 3,
-		},
-	},
-	"qwen3-0.6b": {
-		maxTokens: 8_192,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.3,
-		outputPrice: 1.2,
-		cacheWritesPrice: 0.3,
-		cacheReadsPrice: 1.2,
-		thinkingConfig: {
-			maxBudget: 30_720,
-			outputPrice: 3,
-		},
-	},
-	"qwen2.5-coder-32b-instruct": {
-		maxTokens: 8_192,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.002,
-		outputPrice: 0.006,
-		cacheWritesPrice: 0.002,
-		cacheReadsPrice: 0.006,
-	},
-	"qwen2.5-coder-14b-instruct": {
-		maxTokens: 8_192,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.002,
-		outputPrice: 0.006,
-		cacheWritesPrice: 0.002,
-		cacheReadsPrice: 0.006,
-	},
-	"qwen2.5-coder-7b-instruct": {
-		maxTokens: 8_192,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.001,
-		outputPrice: 0.002,
-		cacheWritesPrice: 0.001,
-		cacheReadsPrice: 0.002,
-	},
-	"qwen2.5-coder-3b-instruct": {
-		maxTokens: 8_192,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.0,
-		outputPrice: 0.0,
-		cacheWritesPrice: 0.0,
-		cacheReadsPrice: 0.0,
-	},
-	"qwen2.5-coder-1.5b-instruct": {
-		maxTokens: 8_192,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.0,
-		outputPrice: 0.0,
-		cacheWritesPrice: 0.0,
-		cacheReadsPrice: 0.0,
-	},
-	"qwen2.5-coder-0.5b-instruct": {
-		maxTokens: 8_192,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.0,
-		outputPrice: 0.0,
-		cacheWritesPrice: 0.0,
-		cacheReadsPrice: 0.0,
-	},
-	"qwen-coder-plus-latest": {
-		maxTokens: 129_024,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 3.5,
-		outputPrice: 7,
-		cacheWritesPrice: 3.5,
-		cacheReadsPrice: 7,
-	},
-	"qwen-plus-latest": {
-		maxTokens: 16_384,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.8,
-		outputPrice: 2,
-		cacheWritesPrice: 0.8,
-		cacheReadsPrice: 2,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 16,
-		},
-	},
-	"qwen-turbo-latest": {
-		maxTokens: 16_384,
-		contextWindow: 1_000_000,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.3,
-		outputPrice: 0.6,
-		cacheWritesPrice: 0.3,
-		cacheReadsPrice: 0.6,
-		thinkingConfig: {
-			maxBudget: 38_912,
-			outputPrice: 6,
-		},
-	},
-	"qwen-max-latest": {
-		maxTokens: 30_720,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 2.4,
-		outputPrice: 9.6,
-		cacheWritesPrice: 2.4,
-		cacheReadsPrice: 9.6,
-	},
-	"qwq-plus-latest": {
-		maxTokens: 8_192,
-		contextWindow: 131_071,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.0,
-		outputPrice: 0.0,
-		cacheWritesPrice: 0.0,
-		cacheReadsPrice: 0.0,
-	},
-	"qwq-plus": {
-		maxTokens: 8_192,
-		contextWindow: 131_071,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.0,
-		outputPrice: 0.0,
-		cacheWritesPrice: 0.0,
-		cacheReadsPrice: 0.0,
-	},
-	"qwen-coder-plus": {
-		maxTokens: 129_024,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 3.5,
-		outputPrice: 7,
-		cacheWritesPrice: 3.5,
-		cacheReadsPrice: 7,
-	},
-	"qwen-plus": {
-		maxTokens: 129_024,
-		contextWindow: 131_072,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.8,
-		outputPrice: 2,
-		cacheWritesPrice: 0.8,
-		cacheReadsPrice: 0.2,
-	},
-	"qwen-turbo": {
-		maxTokens: 1_000_000,
-		contextWindow: 1_000_000,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 0.3,
-		outputPrice: 0.6,
-		cacheWritesPrice: 0.3,
-		cacheReadsPrice: 0.6,
-	},
-	"qwen-max": {
-		maxTokens: 30_720,
-		contextWindow: 32_768,
-		supportsImages: false,
-		supportsPromptCache: false,
-		inputPrice: 2.4,
-		outputPrice: 9.6,
-		cacheWritesPrice: 2.4,
-		cacheReadsPrice: 9.6,
-	},
-	"deepseek-v3": {
-		maxTokens: 8_000,
-		contextWindow: 64_000,
-		supportsImages: false,
-		supportsPromptCache: true,
-		inputPrice: 0,
-		outputPrice: 0.28,
-		cacheWritesPrice: 0.14,
-		cacheReadsPrice: 0.014,
-	},
-	"deepseek-r1": {
-		maxTokens: 8_000,
-		contextWindow: 64_000,
-		supportsImages: false,
-		supportsPromptCache: true,
-		inputPrice: 0,
-		outputPrice: 2.19,
-		cacheWritesPrice: 0.55,
-		cacheReadsPrice: 0.14,
-	},
-	"qwen-vl-max": {
-		maxTokens: 30_720,
-		contextWindow: 32_768,
-		supportsImages: true,
-		supportsPromptCache: false,
-		inputPrice: 3,
-		outputPrice: 9,
-		cacheWritesPrice: 3,
-		cacheReadsPrice: 9,
-	},
-	"qwen-vl-max-latest": {
-		maxTokens: 129_024,
-		contextWindow: 131_072,
-		supportsImages: true,
-		supportsPromptCache: false,
-		inputPrice: 3,
-		outputPrice: 9,
-		cacheWritesPrice: 3,
-		cacheReadsPrice: 9,
-	},
-	"qwen-vl-plus": {
-		maxTokens: 6_000,
-		contextWindow: 8_000,
-		supportsImages: true,
-		supportsPromptCache: false,
-		inputPrice: 1.5,
-		outputPrice: 4.5,
-		cacheWritesPrice: 1.5,
-		cacheReadsPrice: 4.5,
-	},
-	"qwen-vl-plus-latest": {
-		maxTokens: 129_024,
-		contextWindow: 131_072,
-		supportsImages: true,
-		supportsPromptCache: false,
-		inputPrice: 1.5,
-		outputPrice: 4.5,
-		cacheWritesPrice: 1.5,
-		cacheReadsPrice: 4.5,
-	},
-} as const satisfies Record<string, ModelInfo>
-export enum QwenApiRegions {
-	CHINA = "china",
-	INTERNATIONAL = "international",
-}
-export type MainlandQwenModelId = keyof typeof mainlandQwenModels
-export type InternationalQwenModelId = keyof typeof internationalQwenModels
-// Set first model in the list as the default model for each region
-export const internationalQwenDefaultModelId: InternationalQwenModelId = Object.keys(
-	internationalQwenModels,
-)[0] as InternationalQwenModelId
-export const mainlandQwenDefaultModelId: MainlandQwenModelId = Object.keys(mainlandQwenModels)[0] as MainlandQwenModelId
 
 // Doubao
 // https://www.volcengine.com/docs/82379/1298459
